feat: support `getParam` and `getOptionalParam`
Support module level parameters and optional parameters:

```javascript
const userModule = buildModule("MyModule", (m: any) => {
  const myNumber = m.getParam("MyNumber");
  const symbol = m.getOptionalParam("tokenSymbol", "TKN")

  const foo = m.contract("Foo");

  m.call(foo, "incByPositiveNumber", {
    args: [myNumber, symbol],
  });

  return { foo };
});

const deployPromise = this.hre.ignition.deploy(userModule, {
  parameters: {
    MyNumber: 123,
  },
});
```

The same binding and executor as a standard parameter are used with
optional parameters, with an additional default value (undefined
when no default value is given).

As we limit the types of param to `string | number` we can use
`undefined` to mean **no default value** within the param binding.

Minimal validation on the presences of required params and against
duplicate params is included.

Params can be passed at the cli with the `--parameters` flag that
accepts a string that will be parsed to a json object.

More details are included in the `README.md` and module guide.