feat: module contract dep equivalent to module dep
If a contract future returned from a module is passed to a subsequent
deployment builder call, the dependency created will be to the containing
module rather than the vertex itself. This makes the following have the
same execution graph:

```ts
const module = m.useModule(MyModule);

m.contract("Foo", { after: [module] });
```

and

```ts
const { bar } = m.useModule(MyModule);

m.contract("Foo", { after: [bar] });
```

In both cases, the entirety of the module is executed before `Foo` is
deployed.

This is done by having `useModule` switch in proxy futures for the return
results of the module. Each proxy points at both the module and the
underlying future. Any dependency on the proxy is resolved to the module,
but any value is resolved to the underlying future.