refactor: make deploy return type explicit
The return type of deploy on the Ignition wrapper now has an explicit
type.

Note: the return value of serialized result for string and number has
been changed to the value itself. I cannot find a current path that
would lead to this code being used, but I am going to defer the cleanup
to a later refactor.