feat: add state command journal
Every command that updates the Deployment state, _during execution_, is
written to a Journal.

The journal is passed in at `Ignition` creation. In the
**hardhat-plugin**, the journal implementation writes to a file that is
a sibling to the module being deployed.

Supporting refactorings:

* remove `ui` flag infavour of constructor

The `Ignition` object can be setup with a UI renderer. If none is given
a noop renderer is put in its place. The deploy function can then focus
on `deploy` options.

This means the `IgnitionWrapper` creates a new ignition object in each
of its tasks.

* refactor(deploymentState): pull out run logic

Create a shared run command function to dry up the repetition on
conversion to commands.