docs(readme): update toc on documentation (#59)
* docs(readme): update toc on documentation

Add the missing pages into the documentation from the readme.

Add shields to the readme as well.

* docs(readme): rename calling contracts section

* docs(readme): change title of network chain id section

* docs(readme): reorder module in modules section

* docs(readme): move module params section

* docs(readme): renaming resumability to general doc

Renaming from `resumability` to a general `running-a-deployment.md` so
there is a place to gather all the info on running a deployment.

* docs(readme): merge visualization into running

Collapse the vizualization doc into the `running-a-deployment.md` page.

* docs(running-a-deployment): fix link to ens example

* docs(readme): rename testing with HH file

* docs(getting-started-guide): add HH tests section

Show off the ability to use ignition in HH tests.

* docs(modules): update with value passing

* docs(testing): update using in hardhat tests

Update the example (as `simple` no longer exists).

* docs(running-a-deployment): add `deploy` section

Add instructions on running the `deploy` task and its flags to the
`running-a-deployment.md` instructions.

* docs: move config section to running the deploy

* docs: add endowment to module docs

* docs(modules): move create2 under contract deploy

* docs(running): split config code example