feat: read messages from journal
The journal uses ndjson, so read them one line at a time through an
async generator.

test(stateReducer): add tests and simplify

Increase the test coverage over the state reducer before changing the
message structure.

Supporting refactors:

* move reducer types out

* separate out execution state reducer

* deployState - restructure execution state

Use a map of vertex ids to states rather than collections of states.

The intermediary batch results are stored against the vertexes structure
rather than the batch. The batch is reduced down to a list of vertex
ids.

* deployState - remove end batch message

Replace the end batch message updating the state with a check in the
individual vertex update to see if the batch is complete.

We also change `batch` in execution state so that it is nullable, either
befor or after a batch is started/completed. This is a change from the
empty set that was used before.

* change the names of execution messages