feat(cli): show message on hold
Display more details of the on hold vertexes when a deployment stops as
on hold.

Supporting refactors:

* refactor(core): remove journal from tx service

The journaling is happening at the execution graph level for the moment.

* refactor(core): encapsulate providers in services

Seal the usage of providers behind services to ease mocking.

* refactor: externalise services

Move services to be external to ignition core. This is to support
mocking. We should look again at the best api for this.

* test(core): add rerun tests

Component level testing of reruns using an in-memory journal and mocking
out the services.

* fix(plugin): serialization/deserializtion of messages

Support more complex return results from ethers in the messages for the
journal.

We need to be able to serialize/deserialize results with BigNumber and
complex `Error` objects. For now we are only holding on to the message
in the error.

refactor(core): remove old journal files