fix: display unexpected errors
General errors during transform were not being reported back, instead
leaking out to the top level but not being shown.

A top level handler has been added to the deploy process in ignition
core that moves the phase to `failed-unexpectedly`. The cli has been
modified to display errors in this state.

A caveat to the general handling of unexpected errors is that any
exception happening in the validation phase is treated as a validation
error rather than an unexpected error.

The top level ui display component has been refactored to show a type
error if it is not covering all phases.

Fixes #141.