chore: switch to npm
We are moving to `npm` and away from `yarn` across nomic.

At the same time the uniswap example is explicit about which version of
@openzeppelin is used. This allows us to leverage hoisting to share
`node_modules` across all packages.

To leverage npm7 and workspaces in the ci, an additional npm install
step is needed. This can be removed once node 14 is out of Maintenance
LTS.

Resolves #147.