refactor: make graph visiting return results generic
To allow different return results across validation and execution, the
visit result types have to become generic in the returning result type.

The validation result can then return undefined as the result carrier
(as validation doesn't do anyting in the success case) while the
execution result can return the specific types for each execution
vertex.

The payoff here is removing the `undefined as any` in validation.