feat: enhance validation error display
Show the file/line/column against each validation error. The display
allows cmd+click in the vscode terminal.

To support this we capture a potential validation error when the
deployment builder action is called. These potential errors are passed
to the validation phase, one for each vertex. If a validation check
fails we use the pre-captured error so that the stack trace shows the
correct point in the module.

We have to do some magic in IgnitionValidationError. Firstly need to
externally rewrite the stack via a function call `resetStackFrom`, this
is because Hardhat Errors manipulate stack themselves. To allow calling
a custom function on the error, we need to manipulate the prototype
becasue ... javascript.

Resolves #87.