refactor: pass providers to Ignition constructor
In place of services, pass providers then immediately convert them. This
has the advantage of only exporting one set of concepts (i.e. low level
providers) rather than two (providers and higher level services),
particularly as one leverages the other.

The consequence of this approach though is that the providers are too
low level for clean mocking in our core tests. The services are the more
appropriate seam. By switching to providers in the `Ignition`
constructor we lose that injection seam. This has been worked around
with a utility function that sets up Ignition but overrides the services
so that mocked services can continue to be used in tests.

Part of #171.