refactor: use a protected constructor plus factory method
On Ignition within core we expose a public factory method that accepts
providers, but delegates onto a protected constructor that deals in
Services. Our tests use a `TestIgnition` that extends the `Ignition`
object with a public constructor that takes `Services`.

Combined this gives us easy mocking in tests but only expose providers
in the public api.