fix: rework ephemeral deployment to support passed artifacts
This involved reworking the deployment loaders to separate between
storing artifacts from HH/artifact resolver and those passed in.

In Ephemeral mode we ignore the passed artifact but keep the name used
to load via the artifact resolver. In file mode we store the artifact
for later retrieval.

The future id is used to identify the artifact to load now instead of
artifact path. For a future this is recorded in the execution state
under `artifactFutureId` to support futures that resolve their artifacts
based on another future.