chore: re-enable samples
Upgrade the samples to toolbox 3.0.0 so that ethers v6 is used
throughout the monorepo.

This lets us bring back the samples, with the exception of the full
uniswap test as the test depends on the uniswap sdk which is ethers v5
based. That test remains skipped but a simpler uniswap test has been
added that asserts that the deploy has completed.

Includes a minor fix to the basic execution strategy to fix the `to` of
send requests.

Resolves #421.