refactor: rename wipe message
Rename the `WIPE_EXECUTION_STATE` message to `WIPE_APPLY`. This is
because all other message follow a `NOUN_VERB_MESSAGE` structure, so we
should follow it for wipe as well.

The problem with mentioning execution state is it suggests that we are
dealing with a execution state level message. In reality `WIPE` is its
own world.