feat: add `Held` status to exState status
To support execution strategies returning `held` status in the future,
we are adding it into the execution status of exStates.

The return result of a deployment can now include a held property
listing the futures that have been held and the reason for the hold.

This is wired through to the ui as a new sort of Held result instead of
treating it as success or failure.

Resolves #451