fix(visualize): separate argument from li
We use the Argument components both in lists and as standalone values
(`to` in send data for example). The previous Argument component
included a li which doesn't work as well for the standalone values.

The `li` is now added at the list level, while the Argument component
returns either a `p` or `a` depending on the type of argument being
displayed.