chore: update node typing and target
We leverage `cause` on error and `at` on the array object. This is
available on recent versions of the `v16.x` of node. We think this is
acceptable in terms of node 16 support and so are bumping our target to
`ES2022`.

At the same time we are increasing our `@types/node` version to be the
latest in the node 16 line.