remove support for module file names in tasks
We are removing support for invoking a module just by giving its file
name. Instead the tasks will work in terms of a full or relative path to
the module file.

Users prefer this for autocomplete reasons anyway, but it removes
confusion as to whether it is the `ModuleId` or filename that is in use
here.

Resolves #527.