improvement: display warning on ephemeral network
Display a warning message in the `deploy` task. It indicates that
deployment is against an ephemeral network.

The implementation is to add extra information to the start run event,
specifically the `deploymentDir`, which then allows the UI event handler
to determine if the deployment is ephemeral.

This means injecting that information down throught the
`deploy`/`deployer` barrier. This is not ideal, but I think it is
legitimate to assume the UI should have access to all the passed in
config.

Resolves #553.