fix: display of Deployment starting panel on reruns
Our display of the deployment starting panel is dependent on the
`RUN_START` event. This event was being triggered only once for a
deployment, on initialization of the deployment state, on the very first
run of the deployment. The panel was not being displayed on subsequent
runs.

Several moves have been made to fix this:

1. The original run start has been renamed to `DEPLOYMENT_INITIALIZE` and
   only has responsibility for setting the `chainId`. It is only ever run
   as the first message in a new deployment.
2. The execution engine is no longer invoked if there are no batches to
   run (i.e. no uncompleted futures).
3. A new `RUN_START` that is UI only has been added to the deployer.
   If there is a batch to execute, and the execution engine is being
   invoked, we trigger the new `RUN_START`.
4. The deployment initialize event in the state reducer only sets the
   `chainId` (rather than clearing executionStates as well).

The UI for displaying the execution panel has been shifted to the new
run start event.

These changes gives us the correct UI combinations:

- Show Deploying panel on fresh deploy
- Only show complete on rerun with no new futures
- Show Deploying panel on rerun with new futures

Fixes #566.