refactor: remove todo about updates in assertion
Add the future id to the builder (used for checking duplicates across
modules and submodules) out from the assertion method and into each api
action. When a future is added to the module, we consider it safe to add
the future id to the list of previously seen future ids.