feat: add `hardhat-plugin-viem` package
Add a `hardhat-plugin-ethers` package and a `hardhat-plugin-viem`
package. They both extend the existing `hardhat-plugin` package but
they are responsible for adding the `IgnitionHelper` for scripts/tests
that is set on `hre.ignition`.

The integration tests remain against `hardhat-plugin` but have been
rework to be `viem` based. A `TestIgnitionHelper` has been added that
has some `viem` machinery. This is duplicated from `hardhat-viem` rather
than using it directly to avoid the dangers of `hardhat-plugin` having
`hardhat-viem`/`hardhat-plugin-viem` deps.

If a user loads only the `hardhat-ignition` plugin but not ethers/viem
and trys to use `hre.ignition` an error is thrown reporting that they
need one of the ethers/viem plugins.

A viem sample project has been added as well and the CONTRIBUTING.md
guide updated.

Resolves #603, #616

---------

Co-authored-by: John Kane <john@kanej.me>