feat: add validation for send to addresses being addresses
Sharpen the validation so that a string passed as the to address that
does not meet the requirements of being an ethereum address fails
`buildModule` validation.

This required updating the tests.

At the same time a sharper address comparison was added to the send
from/to test - which is what prompted the addition of the more general
address valid address check.