feat: enhance deploy return types using HH viem type generation
Use an enhanced typed for the return result of `hre.ignition.deploy`
that is based on contracts returned from the result.

This is done in two ways:

* pull the type information from HH-viem's code generation during
  compile of the type information for the contract (under artifacts)
* enhance the Ignition `Artifact` type to carry a full Abi type, which
  can then be passed to viem for conversion into a sophisticated
  contract type

The artifact type is now base on `AbiType`, which underlies `viem`.

Because the tests are dependent on artifact metadata generated by Hardhat-viem
on a compile. We now trigger a HH compile on the fixture projects before a build
of the `hardhat-plugin-viem` package.