Added `verify` task (#632)
Adds a verify task to `hardhat-plugin`. Also add a `--verify` parameter
to the `deploy` task, which triggers a verification after a successful deploy.

Only Etherscan is supported.

Verification is done on a deployment (so only file based deployments). It 
reads the deployment state and resolves the `chainid` and the list of 
futures that were successful contract deployments (so no contractAts).

The execution state provides the contructor args/libraries/artifact that
were used to deploy and so are exactly what is needed for verification.

The `verify` task leverages the `hardhat-verify` plugin. It is pulled in as
a peer dependency. The exact same config as is used in `hardhat-verify`
is used for HH Ignition verify.

Resolves #630

---------

Co-authored-by: John Kane <john@kanej.me>
Co-authored-by: Patricio Palladino <email@patriciopalladino.com>