chore: cleanup eslint config in examples
We were seeing errors on some sample projects eslinting in vscode. This was due
to a misconfiguraiont in their eslintrc that was pulling in the tsconfig wrong.
Eslint then fell back to the root repo directories tsconfig, which doesn't
exist - hence the error.

The solution was to remove the failing import and rely on the default.