refactor: limit exposing of strategy types
We are going to limit the types exposed by the two exported strategies.

Our intent to expand release full Strategy support. In the meantime we are
limiting the types exposed. A hack has been put in place that exports reduced
versions of the BasicStrategy and the Create2Strategy, but then sneakly replaces
the implementation with alternate classes. There are simpler hacks, but this one
works with API Extractor.