fix: deal with deployments started in previous versions
A journal from the previous version will have a strategy name but not a config.

That means any deployment restarted with an in flight transaction will fail as
the comparison would have been between undefined and the new basic strategy
config `{}`.

Instead, we now convert `undefined` for strategy config into `{}`. All future
versions of Ignition will have that property defined. We know it is equivalent
to `{}` as only the basic strategy was allowed originally.