fix: use Hardhat 2.20.0 for dev, but keep peer deps the same
The Ignition test suite fails on an EDR bug in `v2.21.0`, for dev dependencies
we are going to use `2.20.0` until the bug is resolved.

The peer dep bound will remain. The test failures are not in Ignition usage, but
in our test setup (which is complicated and generates edge cases that EDR is
resolving).