Fix simulation after gas estimation failure.
If the gas estimation fails, we simulate the transaction so that we
can get access to the return data. During that simulation we weren't
setting the gas limit, so some clients would use the block gas limit,
and most users don't have enough eth to afford that and the simulation
fails.