fix: fqn resolution of artifacts
In the artifact resolver detect whether a FQN is being used to refer to the
contract and use Hardhat's own build info resolution method as we can rely on
the fqn.

This is now covered with an additional integration test for repos with multiple
contracts with the same name in different files and an Ignition module where
those contracts are referred to by FQNs.

Fixes #778.